refactor: extract ProfileCard to component
